metadata: Extract the metadata handlers from netapplier
